Early validation of some JSON Schema errors
This adds a rather crude little parser combinator library that lets
us start catching JSON Schema validation errors earlier the document
than our previous approach did.

It falls very far short of being able to perfectly validate JSON
Schema, or of being able to catch errors at the earliest possible
moment, and doesn't currently handle at least one known case where
the current approach falls down (oneOf constraints), but it does
manage to fix the most egregious case (continuing long string literals
in places where there's no valid way to close them. At least, it does
this in most of the cases we care about).

The big virtue of this approach is that it gives us a very flexible hook
for fixing any particular problems we run into, and a place to add
special cases for schemas that we run into that are problematic.

The big downside is that I had to write a JSON parser and it's not very
pretty.